adds support for cluster node rolling updates
I don't love this solution but it works well enough. CloudFormation supports rolling updates of nodes in an autoscaling group. It can batch changes and add a delay before moving onto new nodes. Terraform does not support this today so I'm using Terraform to manage a CloudFormation stack. With this change, we can update the launch configuration and avoid downtime by replacing the nodes in the autoscaling group one at a time. There is a 3 minute delay to allow time for the new nodes to boot and join the cluster. In the future, we should add a health check to make sure that nodes successfully join the cluster.