Add MongoDB metrics exporter, ServiceMonitor and Prometheus Alert rules.
The metrics exporter is developed by Percona https://github.com/percona/mongodb_exporter
I have created my own Docker container as there were not official ones.

For the moment the only Prometheus rule for alerting is based on number of members in the replica set. We can extend this in the future if we consider it necessary.

I have also created a dashboard in Grafana displaying some of the metrics scraped.