Reflect the current worker node group (#282)
Note that Terraform will complain if trying to apply this. It does not want to create a node group with 0 workers. However removing this will result in a cluster outage. This means that currently Terraform is not usable at all. If you really need to use it, proceed removing the contents of this PR, and run "terraform state rm aws_eks_node_group.nodes"